Minor refactor, documentation and polish
