+++ Assets/Scripts/WorldGen/MeshData.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: c7e0b728a0b17ba489db3d9e82bf7efd
+timeCreated: 1518799453
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
