+++ Assets/Scripts/WorldGen/MeshDataGenerator.cs
@@ -4,61 +4,6 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
-// had to change this to make it more thread compatible.
-
-public class MeshData {
-    public Vector3[] vertices;
-    public int[] triangles;
-    public Color[] colors;
-    public Vector2[] uvs;
-    //public Vector2[] uvs2;
-
-
-    public MeshData[] split() {
-        int maxVertices = 60000;
-        if (vertices.Length <= maxVertices)
-            return new MeshData[]{ this };
-
-        List<MeshData> splitData = new List<MeshData>();
-        int numMeshes = Mathf.CeilToInt(vertices.Length / (float)maxVertices);
-
-        int maxVertsPerMesh = vertices.Length / numMeshes;
-        maxVertsPerMesh = Mathf.CeilToInt(maxVertsPerMesh / 4.0f) * 4;
-        int trianglesPerMesh = maxVertsPerMesh * 3 / 2;
-
-        for(int m = 0; m < numMeshes; m++) {
-            List<Vector3> vertices = new List<Vector3>();
-            List<int> triangles = new List<int>();
-            List<Color> colors = new List<Color>();
-            List<Vector2> uvs = new List<Vector2>();
-            //List<Vector2> uvs2 = new List<Vector2>();
-
-            for (int v = 0; v < maxVertsPerMesh; v++) {
-                if (m * maxVertsPerMesh + v >= this.vertices.Length)
-                    break;
-                vertices.Add(this.vertices[m * maxVertsPerMesh + v]);
-                colors.Add(this.colors[m * maxVertsPerMesh + v]);
-                uvs.Add(this.uvs[m * maxVertsPerMesh + v]);
-                //uvs2.Add(this.uvs2[m * maxVertsPerMesh + v]);
-            }
-            for (int t = 0; t < trianglesPerMesh; t++) {
-                if (m * trianglesPerMesh + t >= this.triangles.Length)
-                    break;
-                triangles.Add(this.triangles[m * trianglesPerMesh + t] - m * maxVertsPerMesh);
-            }
-            MeshData resPart = new MeshData();
-            resPart.vertices = vertices.ToArray();
-            resPart.triangles = triangles.ToArray();
-            resPart.colors = colors.ToArray();
-            resPart.uvs = uvs.ToArray();
-            //resPart.uvs2 = uvs2.ToArray();
-            splitData.Add(resPart);
-
-        }
-        return splitData.ToArray();
-    }
-}
-
 /// <summary>
 /// A Voxel Mesh generator 
 /// </summary>
@@ -69,7 +14,6 @@ public class MeshDataGenerator {
     protected List<int> triangles = new List<int>();
     protected List<Color> colors = new List<Color>();
     protected List<Vector2> uvs = new List<Vector2>();
-    //protected List<Vector2> uvs2 = new List<Vector2>();
     protected BlockData[,,] pointmap;
 
     System.Random rnd = new System.Random(System.DateTime.Now.Millisecond);
@@ -83,6 +27,7 @@ public class MeshDataGenerator {
     }
     protected MeshDataType meshDataType;
 
+
     /// <summary>
     /// NB! Not thread safe! Do not call from threads other then the main thread.
     /// Generates a mesh from MeshData.
@@ -95,17 +40,17 @@ public class MeshDataGenerator {
         mesh.triangles = md.triangles;
         mesh.colors = md.colors;
         mesh.uv = md.uvs;
-        //mesh.uv2 = md.uvs2;
         mesh.RecalculateNormals(); //Normals could be provided by MeshData instead, to save mainthread cpu time.
         return mesh;
     }
+
 
     /// <summary>
     /// Generates all data needed for a mesh of cubes
     /// </summary>
     /// <param name="pointmap">Point data used to build the mesh.
     /// The outermost layer (in x and z) is used to decide whether to add faces on the cubes on the second outermost layer (in x and z).</param>
-    /// <returns>a mesh made from the input data</returns>
+    /// <returns>an array of meshdata objects made from input data</returns>
     public static MeshData[] GenerateMeshData(BlockData[,,] pointmap, float voxelSize = 1f, Vector3 offset = default(Vector3), MeshDataType meshDataType = MeshDataType.TERRAIN) {
         MeshDataGenerator MDG = new MeshDataGenerator();
         MDG.meshDataType = meshDataType;
@@ -126,10 +71,10 @@ public class MeshDataGenerator {
         meshData.triangles = MDG.triangles.ToArray();
         meshData.colors = MDG.colors.ToArray();
         meshData.uvs = MDG.uvs.ToArray();
-        //meshData.uvs2 = MDG.uvs2.ToArray();
 
         return meshData.split();
     }
+
 
     /// <summary>
     /// Generates the mesh data for a cube
@@ -145,6 +90,11 @@ public class MeshDataGenerator {
         if (cubePos.z != 0 && checkIfSolidVoxel(cubePos + new Vector3Int(0, 0, -1)) == false) GenerateCubeFace(FaceDirection.zm, cubePos - offset, blockData, voxelSize);
     }
 
+    /// <summary>
+    /// Checks if a voxels is fully opaque
+    /// </summary>
+    /// <param name="voxelPos">position of voxel</param>
+    /// <returns>Whether the voxel is opaque</returns>
     protected bool checkIfSolidVoxel(Vector3Int voxelPos) {
         if (pointmap[voxelPos.x, voxelPos.y, voxelPos.z].blockType == BlockData.BlockType.NONE ||
             pointmap[voxelPos.x, voxelPos.y, voxelPos.z].blockType == BlockData.BlockType.WATER)
@@ -212,6 +162,7 @@ public class MeshDataGenerator {
 
     }
 
+
     /// <summary>
     /// Stores the indices of the texture slices to use for a face of a block.
     /// </summary>
@@ -239,12 +190,9 @@ public class MeshDataGenerator {
         }
 
         for (int i = 0; i < 4; i++)
-            colors.Add(new Color((int)texTypes[0], (int)texTypes[1], 0));                  // Because Unity does not have an official way of sending 
-                                                                        //  the slice info to the shader we store it in the colour.
-                                                                        //  Basetype in red channel and modifier in green channel.
-                                                                        // If we ever need to use the color channels for something
-                                                                        //  more important, this could probably be sent via uv2 instead.
+            colors.Add(new Color((int)texTypes[0], (int)texTypes[1], 0)); // Using the color to store the texture type of the vertices
     }
+
 
     /// <summary>
     /// Adds texture coordinates for a face of a block.
